API sanctum 추가

+++ app/Http/Controllers/CommentController.php
@@ -17,12 +17,12 @@ class CommentController extends Controller
             return abort(404);
         }
 
+        $user = $request->user();
         $content = $request->input('content');
-        $author =  $request->input('author');
 
         $comment = new Comment();
+        $comment->user_id = $user->id;
         $comment->post_id = $post->id;
-        $comment->author = $author;
         $comment->content = $content;
         $comment->save();
 
@@ -36,6 +36,11 @@ class CommentController extends Controller
         $comment = Comment::where('post_id', $postId)->where('id', $id)->first();
         if (!comment) abort(404);
 
+        $user = $request->user();
+
+        if ($user->id != $comment->user_id) {
+            abort(404);
+        }
         $comment->delete();
 
         return response()->json([
